requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy[asyncio]
pydantic_settings
greenlet
psycopg2-binary
alembic
pyjwt
passlib